moved cron_update endpoint into a fastapi router
